pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opencmo"
version = "0.1.0"
description = "Open-source AI CMO multi-agent system for indie developers"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.12.4",
    "crawl4ai",
    "python-dotenv",
    "httpx>=0.27",
    "aiosqlite>=0.20",
]

[project.optional-dependencies]
browser = ["browser-use", "playwright"]
geo-premium = ["anthropic>=0.49", "google-generativeai>=0.8"]
scheduler = ["apscheduler>=3.11"]
web = ["fastapi>=0.115", "uvicorn>=0.34", "jinja2>=3.1"]
publish = ["praw>=7.7", "tweepy>=4.14"]
tavily = ["tavily-python>=0.5"]
all = ["opencmo[geo-premium,scheduler,web,publish,tavily]"]

[project.scripts]
opencmo = "opencmo.cli:main"
opencmo-web = "opencmo.web.app:run_server"

[tool.pytest.ini_options]
asyncio_mode = "strict"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501", "E402"]
